refactor: 카테고리 스타일 설정 파일 합치기
